Adding the OST-API module with some services
Created the OST-API module to hold the storage services
and the low-level API of the application.

Created the Auth.login Auth.logout Auth.register functions
and adding the documentation and unit-tests for it.

Also created the MemoryAdapter service to communicate with
the browser storage. Created the documentation and the
unit-tests for this service too.